// dashboard/dashboard/src/types/common-types.ts
export enum HomeComponentTypes{
    Dashboard = 'Dashboard',
    Account = 'Account',
    AddProject = 'Add Project',
    AddSkills = 'Add Skills',
    AddApplication = 'Add Application',
    AddTimeline = 'Add Timeline',
    Messages = 'Messages'
}

export enum AccountCommponentTypes{
    Profile = 'Profile',
    UpdateProfile = 'Update Profile',
    UpdatePassword = 'Update Password',
}